drive-bq-indexer/requirements.txt
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
google-cloud-aiplatform==1.70.0
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2
python-pptx==1.0.2
fastapi==0.114.2
uvicorn[standard]==0.30.6
pandas==2.2.2
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.66.5
requests==2.32.3
rank-bm25==0.2.2
numpy==1.26.4
scikit-learn==1.5.2
